Elevator part 2
The elevator now has functions to return floor numbers based on the floor's name or destination map. When using an elevator to move between mods, it'll try to use these to find an appropriate floor to send you to if the number isn't set for whatever reason.

Also fixed that one typo in sans's dialogue.